Extracted out all of the board's responsibilities into separate classes. Allows for much easier testing and dependency injection.
There is no new functionality but most files have changed as a result. The break down of board dependencies are:
- BoardCache
- BoardInitialiser
- CheckManager
- MoveValidator
- Pawn Manager

Consequently all the board needs to do to test the behaviour of these classes is ensure that it calls them. This has been done for everything that exists, except there is still the issue of call order of methods - for example the check states should be updated after the board cache has been, since it uses the cached king squares. This isn't currently covered by tests and will need to be.